+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - **_Breaking Change_**: move `calc_buffer_size` into trait `OutputDescExt`.
 - **_Breaking Change_**: rename methods `get_xxx` to `xxx`.
+- Feat: add `Capturer.buffer_mut`.
 
 ## v0.1.0